Moving ContainerSerializer to its own file
